analytics/ml/requirements.txt
pandas
scikit-learn
sqlalchemy
pymysql
python-dotenv
